use int64 insead of int into values.Int

+++ pkg/html/dynamic/eval/eval.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+
 	"github.com/MontFerret/ferret/pkg/runtime/core"
 	"github.com/MontFerret/ferret/pkg/runtime/values"
 	"github.com/mafredri/cdp"
@@ -68,7 +69,7 @@ func Property(
 	// all props
 	if propName == "" {
 		var arr *values.Array
-		arr = values.NewArray(len(res.Result))
+		arr = values.NewArray(int64(len(res.Result)))
 
 		for _, prop := range res.Result {
 			val, err := Unmarshal(prop.Value)
