+++ pkg/stdlib/datetime/diff.go
@@ -64,7 +64,7 @@ func DateDiff(_ context.Context, args ...core.Value) (core.Value, error) {
 	}
 
 	if !isFloat {
-		return values.NewInt(int(unitDiff)), nil
+		return values.NewInt(int64(unitDiff)), nil
 	}
 
 	return values.NewFloat(unitDiff), nil
