+++ pkg/stdlib/math/ceil.go
@@ -24,5 +24,5 @@ func Ceil(_ context.Context, args ...core.Value) (core.Value, error) {
 		return values.None, err
 	}
 
-	return values.NewInt(int(math.Ceil(toFloat(args[0])))), nil
+	return values.NewInt(int64(math.Ceil(toFloat(args[0])))), nil
 }
