+++ pkg/stdlib/strings/find.go
@@ -28,7 +28,7 @@ func FindFirst(_ context.Context, args ...core.Value) (core.Value, error) {
 	runes := []rune(text)
 	search := args[1].String()
 	start := values.NewInt(0)
-	end := values.NewInt(int(len(text)))
+	end := values.NewInt(int64(len(text)))
 
 	if argsCount == 3 {
 		arg3 := args[2]
@@ -46,10 +46,10 @@ func FindFirst(_ context.Context, args ...core.Value) (core.Value, error) {
 		}
 	}
 
-	found := strings.Index(string(runes[start:end]), search)
+	found := int64(strings.Index(string(runes[start:end]), search))
 
 	if found > -1 {
-		return values.NewInt(found + int(start)), nil
+		return values.NewInt(found + int64(start)), nil
 	}
 
 	return values.NewInt(found), nil
@@ -75,7 +75,7 @@ func FindLast(_ context.Context, args ...core.Value) (core.Value, error) {
 	runes := []rune(text)
 	search := args[1].String()
 	start := values.NewInt(0)
-	end := values.NewInt(int(len(text)))
+	end := values.NewInt(int64(len(text)))
 
 	if argsCount == 3 {
 		arg3 := args[2]
@@ -93,10 +93,10 @@ func FindLast(_ context.Context, args ...core.Value) (core.Value, error) {
 		}
 	}
 
-	found := strings.LastIndex(string(runes[start:end]), search)
+	found := int64(strings.LastIndex(string(runes[start:end]), search))
 
 	if found > -1 {
-		return values.NewInt(found + int(start)), nil
+		return values.NewInt(found + int64(start)), nil
 	}
 
 	return values.NewInt(found), nil
